requirements.txt
alembic==1.11.1
anyio==3.7.1
APScheduler==3.10.4
certifi==2023.7.22
exceptiongroup==1.1.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
python-dotenv==1.0.0
python-telegram-bot==20.4
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
tzlocal==5.0.1
